feat\!: remove wtp completion command
BREAKING CHANGE: The `wtp completion` command has been removed entirely.
Users must now use `wtp shell` for both tab completion and cd functionality.

This simplifies the codebase and provides a cleaner user experience:
- Single command for all shell integration needs
- No confusion between completion vs shell commands
- Reduced maintenance burden (no duplicate code)

Migration guide:
  Before: eval "$(wtp completion zsh)"
  After:  eval "$(wtp shell zsh)"
