+++ cmd/wtp/shell.go
@@ -3,8 +3,14 @@ package main
 import (
 	"context"
 	"fmt"
+	"io"
 	"os"
+	"os/exec"
+	"path/filepath"
+	"strings"
 
+	"github.com/satococoa/wtp/internal/config"
+	"github.com/satococoa/wtp/internal/git"
 	"github.com/urfave/cli/v3"
 )
 
@@ -33,6 +39,77 @@ func NewShellCommand() *cli.Command {
 				Usage:       "Generate fish integration script",
 				Description: "Generate fish integration script with tab completion and cd functionality",
 				Action:      shellFish,
+			},
+			{
+				Name:   "__branches",
+				Hidden: true,
+				Usage:  "List branches for completion",
+				Action: func(_ context.Context, cmd *cli.Command) error {
+					w := cmd.Root().Writer
+					if w == nil {
+						w = os.Stdout
+					}
+					printBranches(w)
+					return nil
+				},
+			},
+			{
+				Name:   "__worktrees",
+				Hidden: true,
+				Usage:  "List worktrees for completion",
+				Action: func(_ context.Context, cmd *cli.Command) error {
+					w := cmd.Root().Writer
+					if w == nil {
+						w = os.Stdout
+					}
+					printWorktrees(w)
+					return nil
+				},
+			},
+			{
+				Name:   "__worktrees_cd",
+				Hidden: true,
+				Usage:  "List worktrees for cd completion with markers",
+				Action: func(_ context.Context, cmd *cli.Command) error {
+					w := cmd.Root().Writer
+					if w == nil {
+						w = os.Stdout
+					}
+
+					// Get current directory
+					cwd, err := os.Getwd()
+					if err != nil {
+						return err
+					}
+
+					// Initialize repository
+					repo, err := git.NewRepository(cwd)
+					if err != nil {
+						return err
+					}
+
+					// Get main worktree path
+					mainRepoPath, err := repo.GetMainWorktreePath()
+					if err != nil {
+						return err
+					}
+
+					// Load config
+					cfg, err := config.LoadConfig(mainRepoPath)
+					if err != nil {
+						return err
+					}
+
+					// Get all worktrees
+					worktrees, err := repo.GetWorktrees()
+					if err != nil {
+						return err
+					}
+
+					// Print with cd-specific formatting
+					printWorktriesForCd(w, worktrees, cwd, cfg, mainRepoPath)
+					return nil
+				},
 			},
 		},
 	}
@@ -66,29 +143,26 @@ _wtp_completion() {
 
     case $cword in
         1)
-            COMPREPLY=( $(compgen -W "add remove list init cd completion shell help" -- "$cur") )
+            COMPREPLY=( $(compgen -W "add remove list init cd shell help" -- "$cur") )
             ;;
         *)
             case "${words[1]}" in
                 add)
                     local branches
-                    branches=$(wtp completion __branches 2>/dev/null)
+                    branches=$(wtp shell __branches 2>/dev/null)
                     COMPREPLY=( $(compgen -W "$branches" -- "$cur") )
                     ;;
                 remove)
                     local worktrees
-                    worktrees=$(wtp completion __worktrees 2>/dev/null)
+                    worktrees=$(wtp shell __worktrees 2>/dev/null)
                     COMPREPLY=( $(compgen -W "$worktrees" -- "$cur") )
-                    ;;
-                completion)
-                    COMPREPLY=( $(compgen -W "bash zsh fish" -- "$cur") )
                     ;;
                 shell)
                     COMPREPLY=( $(compgen -W "bash zsh fish" -- "$cur") )
                     ;;
                 cd)
                     local worktrees
-                    worktrees=$(wtp completion __worktrees_cd 2>/dev/null)
+                    worktrees=$(wtp shell __worktrees_cd 2>/dev/null)
                     COMPREPLY=( $(compgen -W "$worktrees" -- "$cur") )
                     ;;
             esac
@@ -154,10 +228,6 @@ _wtp() {
                         '(--help -h)'{--help,-h}'[Show help]' \
                         '1: :_wtp_worktrees'
                     ;;
-                completion)
-                    _arguments -s \
-                        '1: :_wtp_shells'
-                    ;;
                 shell)
                     _arguments -s \
                         '1: :_wtp_shells'
@@ -179,7 +249,6 @@ _wtp_commands() {
         'list:List all worktrees'
         'init:Initialize configuration file'
         'cd:Change directory to worktree'
-        'completion:Generate shell completion script'
         'shell:Generate shell integration script'
         'help:Show help'
     )
@@ -188,7 +257,7 @@ _wtp_commands() {
 
 _wtp_branches() {
     local branches
-    branches=(${(f)"$(wtp completion __branches 2>/dev/null)"})
+    branches=(${(f)"$(wtp shell __branches 2>/dev/null)"})
     if [[ ${#branches[@]} -gt 0 ]]; then
         _describe 'branches' branches
     fi
@@ -196,7 +265,7 @@ _wtp_branches() {
 
 _wtp_worktrees() {
     local worktrees
-    worktrees=(${(f)"$(wtp completion __worktrees 2>/dev/null)"})
+    worktrees=(${(f)"$(wtp shell __worktrees 2>/dev/null)"})
     if [[ ${#worktrees[@]} -gt 0 ]]; then
         _describe 'worktrees' worktrees
     fi
@@ -204,7 +273,7 @@ _wtp_worktrees() {
 
 _wtp_worktrees_cd() {
     local worktrees
-    worktrees=(${(f)"$(wtp completion __worktrees_cd 2>/dev/null)"})
+    worktrees=(${(f)"$(wtp shell __worktrees_cd 2>/dev/null)"})
     if [[ ${#worktrees[@]} -gt 0 ]]; then
         _describe 'worktrees' worktrees
     fi
@@ -282,4 +351,219 @@ end
 `
 	fmt.Fprint(w, fishIntegration)
 	return nil
+}
+
+// Helper functions from completion.go
+
+// completeBranches provides branch name completion
+func completeBranches(_ context.Context, cmd *cli.Command) {
+	w := cmd.Root().Writer
+	if w == nil {
+		w = os.Stdout
+	}
+	printBranches(w)
+}
+
+// completeWorktrees provides worktree path completion for remove command
+func completeWorktrees(_ context.Context, cmd *cli.Command) {
+	w := cmd.Root().Writer
+	if w == nil {
+		w = os.Stdout
+	}
+	printWorktrees(w)
+}
+
+// isWorktreeManagedCompletion determines if a worktree is managed by wtp (for completion)
+func isWorktreeManagedCompletion(worktreePath string, cfg *config.Config, mainRepoPath string, isMain bool) bool {
+	// Main worktree is always managed
+	if isMain {
+		return true
+	}
+
+	// Get base directory - use default config if config is not available
+	if cfg == nil {
+		// Create default config when none is available
+		defaultCfg := &config.Config{
+			Defaults: config.Defaults{
+				BaseDir: "../worktrees",
+			},
+		}
+		cfg = defaultCfg
+	}
+
+	baseDir := cfg.ResolveWorktreePath(mainRepoPath, "")
+	// Remove trailing slash if it exists
+	baseDir = strings.TrimSuffix(baseDir, "/")
+
+	// Check if worktree path is under base directory
+	absWorktreePath, err := filepath.Abs(worktreePath)
+	if err != nil {
+		return false
+	}
+
+	absBaseDir, err := filepath.Abs(baseDir)
+	if err != nil {
+		return false
+	}
+
+	// Check if worktree is within base directory
+	relPath, err := filepath.Rel(absBaseDir, absWorktreePath)
+	if err != nil {
+		return false
+	}
+
+	// If relative path starts with "..", it's outside base directory
+	return !strings.HasPrefix(relPath, "..")
+}
+
+// printBranches prints available branch names for completion
+func printBranches(w io.Writer) {
+	// Get current directory
+	cwd, err := os.Getwd()
+	if err != nil {
+		return
+	}
+
+	// Get all branches using git for-each-ref for better control
+	gitCmd := exec.Command("git", "for-each-ref", "--format=%(refname:short)", "refs/heads", "refs/remotes")
+	gitCmd.Dir = cwd
+	output, err := gitCmd.Output()
+	if err != nil {
+		return
+	}
+
+	branches := strings.Split(strings.TrimSpace(string(output)), "\n")
+
+	// Use a map to avoid duplicates
+	seen := make(map[string]bool)
+
+	for _, branch := range branches {
+		if branch == "" {
+			continue
+		}
+
+		// Skip HEAD references and bare origin
+		if branch == "origin/HEAD" || branch == "origin" {
+			continue
+		}
+
+		// Remove remote prefix for display, but keep track of what we've seen
+		displayName := branch
+		if strings.HasPrefix(branch, "origin/") {
+			// For remote branches, show without the origin/ prefix
+			displayName = strings.TrimPrefix(branch, "origin/")
+		}
+
+		// Skip if already seen (handles case where local and remote have same name)
+		if seen[displayName] {
+			continue
+		}
+
+		seen[displayName] = true
+		fmt.Fprintln(w, displayName)
+	}
+}
+
+// printWorktrees prints existing worktree names for completion
+func printWorktrees(w io.Writer) {
+	// Get current directory
+	cwd, err := os.Getwd()
+	if err != nil {
+		return
+	}
+
+	// Initialize repository
+	repo, err := git.NewRepository(cwd)
+	if err != nil {
+		return
+	}
+
+	// Get main worktree path
+	mainRepoPath, err := repo.GetMainWorktreePath()
+	if err != nil {
+		return
+	}
+
+	// Load config
+	cfg, err := config.LoadConfig(mainRepoPath)
+	if err != nil {
+		return
+	}
+
+	// Get all worktrees
+	worktrees, err := repo.GetWorktrees()
+	if err != nil {
+		return
+	}
+
+	// For now, keep the old behavior for backward compatibility
+	// This will be called from hidden __worktrees command
+	printWorktriesForRemove(w, worktrees, cfg, mainRepoPath)
+}
+
+// printWorktriesForCd prints worktrees for cd command with special formatting
+func printWorktriesForCd(
+	w io.Writer, worktrees []git.Worktree, currentPath string, cfg *config.Config, mainRepoPath string,
+) {
+	// Print @ for main worktree
+	for i := range worktrees {
+		wt := &worktrees[i]
+		if wt.IsMain {
+			if wt.Path == currentPath {
+				fmt.Fprintln(w, "@*")
+			} else {
+				fmt.Fprintln(w, "@")
+			}
+			break
+		}
+	}
+
+	// Print other worktrees with current marker (managed only)
+	for i := range worktrees {
+		wt := &worktrees[i]
+		if !wt.IsMain && isWorktreeManagedCompletion(wt.Path, cfg, mainRepoPath, wt.IsMain) {
+			name := getWorktreeNameFromPath(wt.Path, cfg, mainRepoPath, wt.IsMain)
+			if wt.Path == currentPath {
+				fmt.Fprintf(w, "%s*\n", name)
+			} else {
+				fmt.Fprintln(w, name)
+			}
+		}
+	}
+}
+
+// getWorktreeNameFromPath calculates the worktree name from its path
+// For main worktree, returns "@"
+// For other worktrees, returns relative path from base_dir
+func getWorktreeNameFromPath(worktreePath string, cfg *config.Config, mainRepoPath string, isMain bool) string {
+	if isMain {
+		return "@"
+	}
+
+	// Get base_dir path
+	baseDir := cfg.Defaults.BaseDir
+	if !filepath.IsAbs(baseDir) {
+		baseDir = filepath.Join(mainRepoPath, baseDir)
+	}
+
+	// Calculate relative path from base_dir
+	relPath, err := filepath.Rel(baseDir, worktreePath)
+	if err != nil {
+		// Fallback to directory name
+		return filepath.Base(worktreePath)
+	}
+
+	return relPath
+}
+
+// printWorktriesForRemove prints worktrees for remove command (no main, no markers, managed only)
+func printWorktriesForRemove(w io.Writer, worktrees []git.Worktree, cfg *config.Config, mainRepoPath string) {
+	for i := range worktrees {
+		wt := &worktrees[i]
+		if !wt.IsMain && isWorktreeManagedCompletion(wt.Path, cfg, mainRepoPath, wt.IsMain) {
+			// Calculate worktree name as relative path from base_dir
+			name := getWorktreeNameFromPath(wt.Path, cfg, mainRepoPath, wt.IsMain)
+			fmt.Fprintln(w, name)
+		}
+	}
 }
